refactored the compute package into several interfaces
