+++ wren/compute/query_context.go
@@ -2,11 +2,9 @@ package compute
 
 import (
 	"context"
-	"time"
 
 	"github.com/google/uuid"
 	"google.golang.org/protobuf/types/known/timestamppb"
-	"google.golang.org/protobuf/types/known/wrapperspb"
 
 	v1alpha "github.com/kaskada-ai/kaskada/gen/proto/go/kaskada/kaskada/v1alpha"
 	"github.com/kaskada-ai/kaskada/wren/ent"
@@ -91,50 +89,4 @@ func (qc *QueryContext) GetComputeTables() []*v1alpha.ComputeTable {
 
 func (qc *QueryContext) Cancelled() bool {
 	return qc.ctx.Err() != nil
-}
-
-func convertKaskadaTableToComputeTable(kaskadaTable *ent.KaskadaTable) *v1alpha.ComputeTable {
-	if kaskadaTable == nil {
-		return nil
-	}
-
-	computeTable := &v1alpha.ComputeTable{
-		Config: &v1alpha.TableConfig{
-			Name:            kaskadaTable.Name,
-			Uuid:            kaskadaTable.ID.String(),
-			TimeColumnName:  kaskadaTable.TimeColumnName,
-			GroupColumnName: kaskadaTable.EntityKeyColumnName,
-			Grouping:        kaskadaTable.GroupingID,
-			Source:          kaskadaTable.Source,
-		},
-		Metadata: &v1alpha.TableMetadata{
-			Schema: kaskadaTable.MergedSchema,
-		},
-		FileSets: []*v1alpha.ComputeTable_FileSet{},
-	}
-
-	if kaskadaTable.SubsortColumnName != nil {
-		computeTable.Config.SubsortColumnName = &wrapperspb.StringValue{Value: *kaskadaTable.SubsortColumnName}
-	}
-	return computeTable
-}
-
-func getComputePreparedFiles(prepareJobs []*ent.PrepareJob) []*v1alpha.PreparedFile {
-	computePreparedFiles := []*v1alpha.PreparedFile{}
-	for _, prepareJob := range prepareJobs {
-		for _, preparedFile := range prepareJob.Edges.PreparedFiles {
-			metadataPath := ""
-			if preparedFile.MetadataPath != nil {
-				metadataPath = *preparedFile.MetadataPath
-			}
-			computePreparedFiles = append(computePreparedFiles, &v1alpha.PreparedFile{
-				Path:         ConvertURIForCompute(preparedFile.Path),
-				MaxEventTime: timestamppb.New(time.Unix(0, preparedFile.MaxEventTime)),
-				MinEventTime: timestamppb.New(time.Unix(0, preparedFile.MinEventTime)),
-				NumRows:      preparedFile.RowCount,
-				MetadataPath: ConvertURIForCompute(metadataPath),
-			})
-		}
-	}
-	return computePreparedFiles
 }
