--- wren/compute/query_result.go
@@ -1,6 +1,0 @@
-package compute
-
-type QueryResult struct {
-	DataTokenId string
-	Paths       []string
-}
